lol
this is such a bad way of doing this. if any window/module has the sub string of a bad app, the game wont run.
This works for most but "ida" is the worst offender, lot of words contain "ida" in it and may appear in apps that are not bad.